+++ src/components/Banner/Banner.jsx
@@ -1,4 +1,5 @@
 import $Banner from './Banner.styles.jsx';
+import Language from './Language/Language.jsx';
 import Logo from './Logo/Logo.jsx';
 import Navigation from './Navigation/Navigation.jsx';
 
@@ -7,6 +8,7 @@ export default function Banner () {
     return (
         <$Banner>
             <Logo />
+            <Language />
             <Navigation />
         </$Banner>
     )