+++ src/components/Main/Home/Home.jsx
@@ -1,57 +1,41 @@
+import { useTranslation } from 'react-i18next';
 import { Link } from "react-router-dom";
 import { $Content, $Home, $Photo, $Span } from "./Home.styles.jsx";
 import MyImage from "../../../assets/my_image_web_no_bg.jpg";
 
 export default function Home() {
+
+  const { t } = useTranslation();
+
   return (
     <$Home>
       <$Content>
-        <$Span $motto={true}>Secure IT for Your Business</$Span>
+        <$Span $motto={true}>{t('motto')}</$Span>
         <p>
-          Athena House specializes in creating secure software with a
-          client-focused approach, utilizing Agile and CI/CD methodologies. My
-          services include coding in Java and JavaScript using various
-          frameworks. Applications are deployed in the cloud, with
-          infrastructure optimized for security.
+          {t('about_athena')}
+          <br />
+          <br />
+          {t('about_developer')}
         </p>
+        <$Span>{t('menu_academy')}</$Span>
         <p>
-          My name is Bartosz Dziurowicz, and I am a Fullstack Developer
-          with many years of experience in the IT industry. I specialize in Java
-          and React, and my skills encompass both backend and frontend
-          development, allowing me to approach projects comprehensively. I am
-          also proficient in cloud technologies, further enhancing my ability to
-          deliver scalable and secure solutions. I hold a Java certification,
-          which confirms my competencies in this field.
+          {t('about_academy')}
+          <Link to="/academy">{t('menu_academy')}</Link>
         </p>
-        <$Span>Academy</$Span>
+        <$Span>{t('menu_software')}</$Span>
         <p>
-          Athena House offers online training courses that are available on the
-          Udemy platform. This allows you to develop your skills at any place
-          and time. Learn from&nbsp;
-          <Link to="/academy">academy</Link>
+          {t('about_software')}
+          <Link to="/software">{t('menu_software')}</Link>
         </p>
-        <$Span>Software</$Span>
+        <$Span>{t('menu_projects')}</$Span>
         <p>
-          Athena House develops unique, original software projects based on my
-          own ideas. Leveraging the latest technologies, I deliver solutions
-          that are secure and tailored to meet specific needs with cloud
-          deployment for optimized security. Check my&nbsp;
-          <Link to="/academy">software</Link>
+          {t('about_projects')}
+          <Link to="/projects">{t('menu_projects')}</Link>
         </p>
-        <$Span>Project</$Span>
+        <$Span>{t('menu_contact')}</$Span>
         <p>
-          The Athena House portfolio showcases diverse projects executed for
-          clients across various industries. Each project highlights my
-          expertise and dedication to providing secure and reliable software
-          solutions tailored to client expectations. View my&nbsp;
-          <Link to="/academy">projects</Link>
-        </p>
-        <$Span>Contact</$Span>
-        <p>
-          Feel free to contact me for any inquiries or collaborations. I'm
-          always ready to discuss your software needs and how Athena House can
-          help you achieve your goals. Let's stay in&nbsp;
-          <Link to="/contact">touch</Link>
+          {t('about_contact')}
+          <Link to="/contact">{t('menu_contact')}</Link>
         </p>
       </$Content>
       <$Photo>